Rename http_protocol package to http_standard
