--- public/dog/1.txt
@@ -1,1 +1,0 @@
-Dog Breed: Corgi
