+++ src/main/java/http_standards/Stringer.java
@@ -1,4 +1,4 @@
-package http_protocol;
+package http_standards;
 
 public class Stringer {
     public static String crlf = "\r\n";
