+++ src/test/java/http_server/RouterTest.java
@@ -1,8 +1,8 @@
 package http_server;
 
-import http_protocol.MIMETypes;
-import http_protocol.RequestCreator;
-import http_protocol.Stringer;
+import http_standards.MIMETypes;
+import http_standards.RequestCreator;
+import http_standards.Stringer;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import org.junit.jupiter.api.BeforeEach;
